- add blog menu auto block
