+++ blocks/blog-listing/blog-listing.css
@@ -1,5 +1,6 @@
 .section.blog-listing-container {
     max-width: 1600px;
+    padding: 0;
 }
 
 .blog-listing.block {
