+++ blocks/blog-menu/blog-menu.css
@@ -1,0 +1,127 @@
+.blog-menu.block {
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+}
+
+.blog-nav {
+    display: flex;
+    z-index: 10;
+    flex-flow: column nowrap;
+    background-color: var(--primary-color);
+    color: var(--white);
+    height: 117px;
+    padding-top: 10px;
+    font-family: var(--font-family-proxima);
+    font-size: var(--body-font-size-m);
+    width: 100vw;
+    max-width: 1600px;
+    box-sizing: border-box;
+    position: sticky;
+    align-self: center;
+    --border-color: var(--white);
+}
+
+.sticky {
+    position: fixed;
+    top: 0;
+    width: 100%;
+}
+
+.blog-nav a  {
+    text-decoration: none;
+    letter-spacing: 0.1em;
+    font-family: var(--font-family-proxima);
+    font-size: var(--body-font-size-m);
+    line-height: 1.5em;
+    color: inherit;
+    text-decoration-style: solid;
+    font-weight: 400;
+}
+
+.blog-nav h1 a {
+    font-size: var(--heading-font-size-xl);
+    font-family: var(--font-family-proxima);
+    font-weight: 700;
+    line-height: 52px;
+    text-align: start;
+    margin-left: 40px;
+    flex-shrink: 10;
+    width: 100%;
+    text-transform: none;
+    letter-spacing: initial;
+}
+
+.blog-nav div {
+    padding: 16px 0;
+    letter-spacing: 0.1em;
+    line-height: 1.5em;
+    background-color: inherit;
+}
+
+.blog-nav ul {
+    visibility: hidden;
+    text-align: center;
+    width: 100%;
+    z-index: 12;
+    background-color: inherit;
+}
+
+.blog-nav li {
+    font-family: var(--font-family-proxima);
+    font-size: var(--body-font-size-m);
+    line-height: 1.4em;
+    letter-spacing: 0.1em;
+    padding: 16px 0;
+    border-top: solid 1px var(--border-color);
+}
+
+.blog-nav .blog-nav-selected {
+    text-align: center;
+}
+
+.blog-nav img {
+    width: auto;
+    vertical-align: middle;
+    filter: brightness(100);
+    padding: 0 0 2px;
+}
+
+.blog-nav .selected-cat {
+    font-weight: bold;
+    border-bottom: 2px solid var(--border-color);
+}
+
+@media (min-width: 992px) {
+    .blog-nav ul {
+        visibility: visible;
+        justify-content: center;
+        display: flex;
+        margin-bottom: 25px;
+    }
+
+    .blog-nav div {
+        display: none;
+    }
+
+    .blog-nav li {
+        border-top: initial;
+        padding: 4px 10px 0 25px;
+    }
+
+    .blog-nav h1 a {
+        margin-left: 66px;
+    }
+}
+
+@media (min-width: 1025px) {
+    .blog-nav h1 a {
+        margin-left: 119px;
+        line-height: 88px;
+        font-size: 3rem;
+    }
+
+    .blog-nav {
+        height: 150px;
+    }
+}
