+++ scripts/scripts.js
@@ -52,6 +52,28 @@ function buildHeroBlock(main) {
   }
 }
 
+function buildBlogNav(main) {
+  const blogTemplates = ['blog-landing-template', 'blog-detail-template'];
+  const template = getMetadata('template');
+  if (blogTemplates.includes(template)) {
+    const h1 = main.querySelector('div:first-of-type > h1');
+    const nav = main.querySelector('div:first-of-type > ul');
+    if (h1 && nav && (nav.compareDocumentPosition(h1) & Node.DOCUMENT_POSITION_PRECEDING)) {
+      const section = document.createElement('div');
+      section.append(buildBlock('blog-menu', { elems: [h1, nav]}));
+      main.prepend(section);
+    }
+  }
+}
+
+function buildBlogDetails(main) {
+  if (getMetadata('template') === 'blog-detail-template') {
+    const section = document.createElement('div');
+    section.append(buildBlock('blog-detail', {elems: []}));
+    main.append(section);
+  }
+}
+
 function buildLiveByMetadata(main) {
   const community = getMetadata('liveby-community');
   if (community) {
@@ -71,6 +93,8 @@ function buildAutoBlocks(main) {
   try {
     buildHeroBlock(main);
     buildLiveByMetadata(main);
+    buildBlogDetails(main);
+    buildBlogNav(main);
   } catch (error) {
     // eslint-disable-next-line no-console
     console.error('Auto Blocking failed', error);
@@ -119,6 +143,7 @@ export function decorateMain(main) {
   decorateSections(main);
   decorateBlocks(main);
   buildFloatingImages(main);
+
 }
 
 /**
